# openapi.yaml
openapi: 3.0.3
info:
  title: Oblixa route contract
  version: 0.1.0
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sb-access-token
    CronSecret:
      type: http
      scheme: bearer
    WebhookSignature:
      type: apiKey
      in: header
      name: stripe-signature
paths:
...
  /api/export/reports:
    get:
      operationId: get_export_reports
      summary: GET /api/export/reports
      description: Generated route contract for src/app/api/export/reports/route.ts.
      security:
        - SessionCookie: []
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "429":
          description: Rate limited
        "500":
          description: Unexpected server error
...
  /api/integrations/calendar/sync:
    get:
      operationId: get_integrations_calendar_sync
      summary: GET /api/integrations/calendar/sync
      description: Generated route contract for src/app/api/integrations/calendar/sync/route.ts.
      security:
        - SessionCookie: []
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "429":
          description: Rate limited
        "500":
          description: Unexpected server error
...